# cmd/api/docs/swagger.yaml
definitions:
  cmd_api_adapters_handlers.TokenResponse:
    properties:
      circulatingSupply:
        items:
          $ref: '#/definitions/domain.CirculatingSupplyEntry'
        type: array
      createdAt:
        type: string
      decimals:
        type: integer
      ercStandard:
        type: string
      frozenChainIds:
        items:
          type: string
        type: array
      issuerId:
        type: string
      metadataURL:
        type: string
      name:
        type: string
      resourceId:
        type: string
      status:
        type: string
      symbol:
        type: string
      totalSupply:
        type: number
      updatedAt:
        type: string
    type: object
  domain.Balance:
    properties:
      amount:
        type: number
      chainId:
        type: string
      createdAt:
        type: string
      ercId:
        $ref: '#/definitions/domain.BigInt'
      id:
        type: string
      resourceId:
        type: string
      updatedAt:
        type: string
    type: object
  domain.BigInt:
    type: object
  domain.CirculatingSupplyEntry:
    properties:
      balance:
        type: number
      participantId:
        type: string
      tokenId:
        description: Only present for non-fungible tokens (ERC721, ERC1155)
        type: string
    type: object
...
  dto.TokenRegistryStatusDto:
    properties:
      createdAt:
        type: string
      decimals:
        type: integer
      ercStandard:
        type: integer
      issuerId:
        type: string
      metadataUrl:
        type: string
      name:
        type: string
      resourceId:
        type: string
      status:
        type: integer
      symbol:
        type: string
      updatedAt:
        type: string
    type: object
...
  governance-api-go_types.Paginated-cmd_api_adapters_handlers_TokenResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/cmd_api_adapters_handlers.TokenResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /audit/tokens:
    get:
      description: Retrieves a paginated list of tokens based on specified query parameters
        for auditing. Returns an empty array if no tokens match the filters.
      parameters:
      - description: 'Items per page (Default: 10, Max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Page number (Default: 1)'
        in: query
        name: page
        type: integer
      - description: Token's name (partial match)
        in: query
        name: name
        type: string
      - description: Issuer ID
        in: query
        name: issuerId
        type: string
      - description: Status of the token.
        enum:
        - new
        - active
        - inactive
        in: query
        name: status
        type: string
      - description: Token type.
        enum:
        - erc20
        - erc721
        - erc1155
        - enygma
        - custom message
        in: query
        name: ercStandard
        type: string
      - description: Symbol of the token
        in: query
        name: symbol
        type: string
      - description: Decimals of the token
        in: query
        name: decimals
        type: integer
      - description: Creation date after which tokens were added (inclusive). Accepts
          Unix ts, YYYY-MM-DD or ISO8601 formats
        in: query
        name: createdAfter
        type: string
      - description: Creation date before which tokens were added (exclusive). Accepts
          Unix ts, YYYY-MM-DD or ISO8601 formats
        in: query
        name: createdBefore
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/governance-api-go_types.Paginated-cmd_api_adapters_handlers_TokenResponse'
        "400":
          description: Invalid parameters specified
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Database error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List Tokens by Query Parameters
      tags:
      - Audit
  /audit/tokens/{resourceId}:
    get:
      description: Retrieves a token by its resource ID
      parameters:
      - description: The resource ID of the token
        in: path
        name: resourceId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api_adapters_handlers.TokenResponse'
        "400":
          description: Invalid resourceId specified
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No tokens found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Database error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Token by Resource ID
      tags:
      - Audit
...
  /resource_info_all_chains/{resourceid}:
    get:
      consumes:
      - application/json
      description: Retrieves balance information for a resource across all chains
      parameters:
      - description: Resource ID
        in: path
        name: resourceid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Array of balances across chains
          schema:
            items:
              $ref: '#/definitions/domain.Balance'
            type: array
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Resource not found in any chain
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get Balance Across All Chains
      tags:
      - Balances
  /resource_info_list_chains:
    post:
      consumes:
      - application/json
      description: Retrieves balance information for a resource across specific chains
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          properties:
            chains:
              items:
                type: string
              type: array
            resource_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Array of balances across specified chains
          schema:
            items:
              $ref: '#/definitions/domain.Balance'
            type: array
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Resource not found in specified chains
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get Balance Across Specific Chains
      tags:
      - Balances
  /resources/{chainid}/{resourceid}:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves all balances for a chain or specific resource balance
        If resourceid is "/" returns all balances, otherwise returns specific resource
      parameters:
      - description: Chain ID
        in: path
        name: chainid
        required: true
        type: string
      - description: Resource ID (use / for all)
        in: path
        name: resourceid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Array of balances or single balance object
          schema: {}
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get Balances in Chain
      tags:
      - Balances
  /token_status/{resourceid}:
    get:
      consumes:
      - application/json
      description: Retrieves token status from the database
      parameters:
      - description: Resource ID
        in: path
        name: resourceid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TokenRegistryStatusDto'
        "400":
          description: Invalid resource ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Token not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get Token Registry Status
      tags:
      - Blockchain
swagger: "2.0"